add more test and reorganize folders
