+++ src/test/scala/com/stratio/khermes/cluster/supervisor/stream/FileSourceTest.scala
@@ -17,7 +17,8 @@ import akka.actor.Props
 import akka.stream.{ActorMaterializer, KillSwitches}
 import akka.stream.scaladsl.{Keep, Sink, Source}
 import com.stratio.khermes.cluster.BaseActorTest
-import com.stratio.khermes.cluster.supervisor.{SourceImplementations, StreamFileOperations, StreamGenericOperations}
+import com.stratio.khermes.cluster.supervisor.SourceImplementations
+import com.stratio.khermes.cluster.supervisor.stream.fileimpl.StreamFileOperations
 import com.stratio.khermes.commons.config.{AppConfig, AppConfigTest}
 import com.stratio.khermes.helpers.twirl.TwirlActorCache
 import com.stratio.khermes.persistence.file.FileClient
